Update and rename use-mobile.tsx to use_mobile.tsx
Debouncing Resize Events: This improves performance by reducing the number of state updates during rapid resize events.

Handling Server-Side Rendering (SSR): Ensures compatibility with SSR frameworks like Next.jsby checking if the window object is defined.

Returning Additional Information: Now also returns the current screen width for more flexibility in your components.
